Move old azure provider
