+++ provider/azure/internal/legacy/azure/azure_test.go
