+++ provider/azure/internal/legacy/azure/config.go
