+++ provider/azure/internal/legacy/azure/config_test.go
