+++ provider/azure/internal/legacy/azure/disks.go
