+++ provider/azure/internal/legacy/azure/disks_test.go
