+++ provider/azure/internal/legacy/azure/environ.go
