+++ provider/azure/internal/legacy/azure/environ_test.go
