+++ provider/azure/internal/legacy/azure/environprovider.go
