+++ provider/azure/internal/legacy/azure/environprovider_test.go
