+++ provider/azure/internal/legacy/azure/export_test.go
