+++ provider/azure/internal/legacy/azure/instance.go
