+++ provider/azure/internal/legacy/azure/instance_test.go
