+++ provider/azure/internal/legacy/azure/instancetype.go
