+++ provider/azure/internal/legacy/azure/instancetype_test.go
