+++ provider/azure/internal/legacy/azure/storage.go
