+++ provider/azure/internal/legacy/azure/storage_test.go
