+++ provider/azure/internal/legacy/azure/userdata.go
